Added decorator state for easier state composition.
Could be a really bad idea, but I'm liking it; there WAS an issue with
inheritance when a class would gain unwanted functionality just because
the parents were poorly chosen. Also tried to be a bit functional and have
player classes define a predicate to tell which units they control.

Kinda bad update but whatever.