States are now gameobjects!
Holy cow! Why do this? Well first of all it seems to have better performance.
The wiring is a bit weird but the good thing about all of this is that states
can now behave nicely with the unity eventsystem. Also, doing it in this way allows
for states to have compositional functionality, which may not be too big of a deal.
The typical use case was, some states would move the cursor and I wanted to stop writing
that for every state that would do it. So now, it's easy enough to just add that component
to provide that functionality instead of making sure it appears in code.