Trying to move away from using game objects to manage state.
Don't know what to say about this milestone. There are HUGE issues on
"Who should respond to input?". The current idea is leveraging the
eventsystem selection to do that with the corresponding interfaces.

Obviously, the issues are that it is used for the GUI stuff, so you have
buttons that get selected that now have no way of getting back to the
system.

While I haven't implemented it, I think all GUI stuff should just have
a new component that reroutes the external buttons to battlesystem.

I dunno man, this has gotten pretty crazy.