Yikes, big commit
Changes:

PlayerTurnController and AI controller about to be scrapped, what these classes
were doing was handling input, which is not really the purpose of their existence;
which was to manage that players resources.

Input is now redirected to the battlesystem, which delgates that input to internally
created state objects. I don't think it was wise to realy on Unity for component based
mixin functionality, but eh.

Also touch functionality has improved, Unity seems more built for THAT than anything else.