Finalish Commit?
Not saying the project is dead, but it's a bit boring for me now. The game
works wellish for a mobile transitioned game; but there were reasons why games
of this genre for mobile were designed to NOT have a huge map. Maybe I shouldn't have made
it mobile, but unity kinda sucks for keyboard input.

Anyway, the game is playable-ish and is pretty functional while being barebones;
I'm proud of it and I learnt a lot so it's all good!